requirements.txt
django>=1.11
django-shopify-webhook >=0.2.6
ShopifyAPI >=2.1.1
jsonfield >=0.9.22
